+++ finance/workflows/src/main/resources/net/corda/finance/workflows/utils/LondonHolidayCalendar.txt
