+++ samples/irs-demo/cordapp/workflows-irs/build.gradle
@@ -14,6 +14,8 @@ dependencies {
 
     // Corda integration dependencies
     cordaCompile project(':core')
+
+    compile("com.fasterxml.jackson.module:jackson-module-kotlin:$jackson_version")
     
     // only included to control the `DemoClock` as part of the demo application
     // normally `:node` should not be depended on in any CorDapps
