+++ samples/irs-demo/cordapp/workflows-irs/src/main/kotlin/net.corda.irs/flows/plugin/FinanceJSONSupport.kt
@@ -1,6 +1,6 @@
 @file:JvmName("FinanceJSONSupport")
 
-package net.corda.irs.utilities
+package net.corda.irs.flows.plugin
 
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.JsonParseException
@@ -10,6 +10,8 @@ import com.fasterxml.jackson.databind.deser.std.StringArrayDeserializer
 import com.fasterxml.jackson.databind.module.SimpleModule
 import net.corda.finance.contracts.BusinessCalendar
 import net.corda.finance.contracts.Expression
+import net.corda.finance.workflows.utils.TEST_CALENDAR_NAMES
+import net.corda.finance.workflows.utils.loadTestCalendar
 import java.time.LocalDate
 import java.util.*
